Changing the option_id from u8 to i32, in order for mongo to work

+++ src/models/result.rs
@@ -2,7 +2,7 @@ use serde::Serialize;
 
 #[derive(Serialize)]
 pub struct ScrutinyResult {
-    pub option_id: u8,
+    pub option_id: i32,
     pub result: i32,
 }
